+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 4.2.0
+## 4.2.0-beta.2
+
+### Added
+- Added `sprinkle:list` bakery command
+
+## 4.2.0-beta.1
 
 ### Changed Requirements
 - Changed minimum Node.js version to **v10.12.0**